+++ _unittests/ut_onnxrt/test_shape_object.py
@@ -238,7 +238,7 @@ class TestShapeObject(ExtTestCase):
 
     def common_test_onnxt_runtime_binary(self, onnx_cl, np_fct,
                                          dtype=numpy.float32):
-        idi = numpy.identity(2)
+        idi = numpy.identity(2, dtype=dtype)
         onx = onnx_cl('X', idi, output_names=['Y'],
                       op_version=get_opset_number_from_onnx())
         X = numpy.array([[1, 2], [3, -4]], dtype=numpy.float64)
